# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitzwilliam Museum Collection Database API",
        "description": "Api for the Fitzwilliam Museum Collection Database",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/agents": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve agents used in the database",
                "description": "A list of agents used in the database, with pagination.",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Aurelius"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "nullable": true
                        },
                        "example": "1"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of the page response",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "nullable": true
                        },
                        "example": "20"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Fields available to return",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "admin.created,admin.modified,admin.id,name,summary_title",
                            "nullable": true
                        },
                        "example": "admin.created,admin.modified,admin.id,name,summary_title"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "enum",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "nullable": true
                        },
                        "example": "asc"
                    },
                    {
                        "name": "sort_field",
                        "in": "query",
                        "description": "Sort field",
                        "required": false,
                        "schema": {
                            "type": "enum",
                            "default": "admin.id",
                            "enum": [
                                "created",
                                "updated",
                                "id",
                                "name",
                                "summary_title"
                            ],
                            "nullable": true
                        },
                        "example": "id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request completed successfully."
                    },
                    "400": {
                        "description": "The request cannot be processed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/agents/{agent}": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve an agent",
                "description": "An agent's details",
                "parameters": [
                    {
                        "name": "agent",
                        "in": "path",
                        "description": "Agent term id number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "agent-28"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Fields available to return",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "admin.id",
                            "nullable": true
                        },
                        "example": "admin.created,admin.modified,admin.id,name,summary_title"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request completed successfully."
                    },
                    "400": {
                        "description": "The request cannot be processed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/auth/signup": {
            "post": {
                "tags": [
                    "Authorisation"
                ],
                "summary": "Sign up a new user",
                "description": "Create a user account",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "Dendera"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    }
}